pldm: Remove unnecessary type casting in logs
This commit removes the unnecessary type casting of values of the
journal logs.

+++ host-bmc/host_pdr_handler.cpp
@@ -723,7 +723,7 @@ void HostPDRHandler::setHostFirmwareCondition()
             return;
         }
         info("Getting the response code '{RC}'", "RC", lg2::hex,
-             static_cast<uint16_t>(response->payload[0]));
+             response->payload[0]);
         this->responseReceived = true;
         getHostPDR();
     };
